refactor to isolate per-endpoint integration tests
